requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
schedule>=1.2.0
python-dateutil>=2.8.0
colorama>=0.4.6
